major reorganization
I'm starting to feel better about this organization but it's still not
perfect.  The kludgey passing of the collection URL to the Race types
doesn't feel clean enough yet.

Everything is starting to shape up nicely and the representations are
starting to seem more reasonable.  Of course, we still need to finish
linking everything together and ensure that the TODO items don't become
larger issues.